Rename the mpg helper package to mpgutil
