+++ internal/command/launch/webui.go
@@ -17,7 +17,7 @@ import (
 	"github.com/superfly/flyctl/helpers"
 	"github.com/superfly/flyctl/internal/command/launch/plan"
 	"github.com/superfly/flyctl/internal/logger"
-	mpgapi "github.com/superfly/flyctl/internal/mpg"
+	"github.com/superfly/flyctl/internal/mpgutil"
 	state2 "github.com/superfly/flyctl/internal/state"
 	"github.com/superfly/flyctl/internal/tracing"
 	"github.com/superfly/flyctl/iostreams"
@@ -102,13 +102,13 @@ func (state *launchState) EditInWebUi(ctx context.Context) error {
 			}
 
 			// Check if region is supported for managed Postgres
-			validRegion, err := mpgapi.IsValidRegion(ctx, region)
+			validRegion, err := mpgutil.IsValidRegion(ctx, region)
 			if err != nil {
 				return fmt.Errorf("failed to validate MPG region: %w", err)
 			}
 
 			if !validRegion {
-				availableCodes, _ := mpgapi.AvailableRegionCodes(ctx)
+				availableCodes, _ := mpgutil.AvailableRegionCodes(ctx)
 
 				return fmt.Errorf("region %s is not available for Managed Postgres. Available regions: %v", region, availableCodes)
 			}
